+++ backend/src/main/java/pw/edu/pl/backend/repository/PlayRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pw.edu.pl.backend.entity.PlayEn;
 
 @Repository
-public interface PlayRepository extends JpaRepository<PlayEn, Integer> {
+public interface PlayRepository extends JpaRepository<PlayEn, Long> {
 }
